add guard authorization API

+++ app/category/controller.js
@@ -2,6 +2,15 @@ const Category = require("./model");
 
 const store = async (req, res, next) => {
   try {
+    //--- cek policy ---/
+    let policy = policyFor(req.user);
+    if (!policy.can("create", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat kategori`,
+      });
+    }
+
     let payload = req.body;
     let category = new Category(payload);
     await category.save();
@@ -22,6 +31,14 @@ const store = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("update", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate kategori`,
+      });
+    }
+
     let payload = req.body;
     let category = await Category.findOneAndUpdate(
       { _id: req.params.id },
@@ -43,6 +60,13 @@ const update = async (req, res, next) => {
 
 const destroy = async (req, res, next) => {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("delete", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus kategori`,
+      });
+    }
     // (1) cari dan hapus categori di MongoDB berdasarkan field _id
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
     // (2) respon ke client dengan data category yang baru saja dihapus;
